requirements.txt
pyserial
python-escpos[all]